+++ include/beast/websocket/detail/decorator.hpp
@@ -29,154 +29,135 @@ struct abstract_decorator
     ~abstract_decorator() = default;
 
     virtual
-    abstract_decorator*
-    copy() = 0;
+    void
+    operator()(request_type& req) const = 0;
 
     virtual
     void
-    operator()(request_type& req) = 0;
-
-    virtual
-    void
-    operator()(response_type& res) = 0;
+    operator()(response_type& res) const = 0;
 };
 
-template<class T>
+template<class F>
 class decorator : public abstract_decorator
 {
-    T t_;
+    F f_;
 
     class call_req_possible
     {
         template<class U, class R = decltype(
-            std::declval<U>().operator()(
+            std::declval<U const>().operator()(
                 std::declval<request_type&>()),
                     std::true_type{})>
         static R check(int);
         template<class>
         static std::false_type check(...);
     public:
-        using type = decltype(check<T>(0));
+        using type = decltype(check<F>(0));
     };
 
     class call_res_possible
     {
         template<class U, class R = decltype(
-            std::declval<U>().operator()(
+            std::declval<U const>().operator()(
                 std::declval<response_type&>()),
                     std::true_type{})>
         static R check(int);
         template<class>
         static std::false_type check(...);
     public:
-        using type = decltype(check<T>(0));
+        using type = decltype(check<F>(0));
     };
 
 public:
-    decorator() = default;
-    decorator(decorator const&) = default;
-
-    decorator(T&& t)
-        : t_(std::move(t))
+    decorator(F&& t)
+        : f_(std::move(t))
     {
     }
 
-    decorator(T const& t)
-        : t_(t)
+    decorator(F const& t)
+        : f_(t)
     {
-    }
-
-    abstract_decorator*
-    copy() override
-    {
-        return new decorator(*this);
     }
 
     void
-    operator()(request_type& req) override
+    operator()(request_type& req) const override
     {
         (*this)(req, typename call_req_possible::type{});
     }
 
     void
-    operator()(response_type& res) override
+    operator()(response_type& res) const override
     {
         (*this)(res, typename call_res_possible::type{});
     }
 
 private:
     void
-    operator()(request_type& req, std::true_type)
+    operator()(request_type& req, std::true_type) const
     {
-        t_(req);
+        f_(req);
     }
 
     void
-    operator()(request_type& req, std::false_type)
+    operator()(request_type& req, std::false_type) const
     {
         req.fields.replace("User-Agent",
             std::string{"Beast/"} + BEAST_VERSION_STRING);
     }
 
     void
-    operator()(response_type& res, std::true_type)
+    operator()(response_type& res, std::true_type) const
     {
-        t_(res);
+        f_(res);
     }
 
     void
-    operator()(response_type& res, std::false_type)
+    operator()(response_type& res, std::false_type) const
     {
         res.fields.replace("Server",
             std::string{"Beast/"} + BEAST_VERSION_STRING);
     }
 };
 
-struct default_decorator
-{
-};
-
 class decorator_type
 {
-    std::unique_ptr<abstract_decorator> p_;
+    std::shared_ptr<abstract_decorator> p_;
 
 public:
+    decorator_type() = delete;
     decorator_type(decorator_type&&) = default;
+    decorator_type(decorator_type const&) = default;
     decorator_type& operator=(decorator_type&&) = default;
+    decorator_type& operator=(decorator_type const&) = default;
 
-    decorator_type(decorator_type const& other)
-        : p_(other.p_->copy())
-    {
-    }
-
-    decorator_type&
-    operator=(decorator_type const& other)
-    {
-        p_ = std::unique_ptr<
-            abstract_decorator>{other.p_->copy()};
-        return *this;
-    }
-
-    template<class T, class =
+    template<class F, class =
         typename std::enable_if<! std::is_same<
-            typename std::decay<T>::type,
+            typename std::decay<F>::type,
                 decorator_type>::value>>
-    decorator_type(T&& t)
-        : p_(new decorator<T>{std::forward<T>(t)})
+    decorator_type(F&& f)
+        : p_(std::make_shared<decorator<F>>(
+            std::forward<F>(f)))
     {
+        BOOST_ASSERT(p_);
     }
 
     void
     operator()(request_type& req)
     {
         (*p_)(req);
+        BOOST_ASSERT(p_);
     }
 
     void
     operator()(response_type& res)
     {
         (*p_)(res);
+        BOOST_ASSERT(p_);
     }
+};
+
+struct default_decorator
+{
 };
 
 } // detail