+++ test/websocket/websocket_echo.cpp
@@ -12,18 +12,23 @@
 
 int main()
 {
+    using namespace beast::websocket;
     using endpoint_type = boost::asio::ip::tcp::endpoint;
     using address_type = boost::asio::ip::address;
 
     try
     {
-        boost::system::error_code ec;
-    	beast::websocket::async_echo_server s1{nullptr, 1};
-        s1.open(true, endpoint_type{
+        beast::error_code ec;
+    	async_echo_server s1{nullptr, 1};
+        s1.open(endpoint_type{
             address_type::from_string("127.0.0.1"), 6000 }, ec);
+        s1.set_option(read_message_max{64 * 1024 * 1024});
+        s1.set_option(auto_fragment{false});
+        //s1.set_option(write_buffer_size{64 * 1024});
 
-    	beast::websocket::sync_echo_server s2(true, endpoint_type{
+    	beast::websocket::sync_echo_server s2(&std::cout, endpoint_type{
 	        address_type::from_string("127.0.0.1"), 6001 });
+        s2.set_option(read_message_max{64 * 1024 * 1024});
 
     	beast::test::sig_wait();
     }