+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fixes for WebSocket echo server
 * Fix 32-bit arm7 warnings
 * Remove unnecessary include
+* WebSocket server examples and test tidying
 
 API Changes: